flash_backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
